requirements.txt
python-dotenv

scipy==1.11.3
scikit-learn==1.2.2
statsmodels==0.14.0

tqdm==4.66.1
eli5==0.13.0
Jinja2==3.1.4
numpy==1.23.5

pandas==1.5.3
pandasai==2.3.0
PyYAML

pygwalker
pycaret
matplotlib==3.7.1
seaborn==0.12.2

xgboost
geopy

dabl
pyjanitor
shap
kneed
